svc-catalogo/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.6.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
minio==7.2.5
pytest==8.1.1